# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Routines & Alerts API
  version: 1.0.1
  description: >
    API RESTful per la gestione di Routine e Alert.
    - Una **Routine** è l'entità principale.
    - Una **Routine** contiene una collezione di **Alert**.
    - Un **Alert** può essere riutilizzato in più Routine.
    - Una Routine senza Alert non è significativa.
servers:
  - url: http://localhost:8080/api
    description: Local server

paths:

  /routines:
    get:
      summary: recupera tutte le Routine
      description: restituisce tutte le routine registrate
      operationId: getAllRoutines
      tags: [Routine]
      responses:
        "200":
          description: lista di tutte le routine
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoutineOutput"
        "500":
          description: errore interno del server
    post:
      summary: crea una Routine
      description: crea una nuova Routine contenente il body passato
      tags: [Routine]
      operationId: createRoutine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineInput"
            example:
              nomeRoutine: "mattina-lavoro"
              alerts:
                - oraInizio: "06:30"
                  oraFine: "06:45"
                  citta: "Silvi"
                  tipoMessaggio: "Sveglia mattutina"
                  codici:
                    - 1001
                - oraInizio: "07:00"
                  oraFine: "07:05"
                  citta: "Silvi"
                  tipoMessaggio: "Promemoria colazione"
                  codici:
                    - 1002
      responses:
        "201":
          description: routine creata con successo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutineOutput"
        "400":
          description: invalid input
        "500":
          description: errore interno del server

  /routines/{id}:
    get:
      summary: recupera una Routine
      description: recupera una Routine univoca tramite il suo id specificato
      tags: [Routine]
      operationId: getRoutineById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID della routine
      responses:
        "200":
          description: singola routine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutineOutput"
        "404":
          description: routine not found
        "500":
          description: errore interno del server
    delete:
      summary: elimina una Routine
      description: elimina la Routine avente l'id specificato
      tags: [Routine]
      operationId: deleteRoutine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID della routine
      responses:
        "204":
          description: eliminazione avvenuta con successo
        "404":
          description: routine not found
        "500":
          description: errore interno del server
    patch:
      summary: patch Routine
      description: aggiorna parzialmente la Routine avente id passato col body della request
      tags: [Routine]
      operationId: patchRoutine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID della routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineInput"
            example:
              nomeRoutine: "mattina-lavoro-aggiornata"
              alerts:
                - oraInizio: "06:35"
                  oraFine: "06:50"
                  citta: "Silvi"
                  tipoMessaggio: "Sveglia aggiornata"
                  codici:
                    - 1003
      responses:
        "200":
          description: routine aggiornata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutineOutput"
        "400":
          description: invalid input
        "404":
          description: routine not found
        "500":
          description: errore interno del server
    put:
      summary: replace Routine
      description: sostituisce completamente la Routine avente id passato, se non esiste la crea
      tags: [Routine]
      operationId: updateRoutine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID della routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineInput"
            example:
              nomeRoutine: "mattina-lavoro-completa"
              alerts:
                - oraInizio: "06:30"
                  oraFine: "06:45"
                  citta: "Silvi"
                  tipoMessaggio: "Sveglia mattutina"
                  codici:
                    - 1001
      responses:
        "200":
          description: routine sostituita
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutineOutput"
        "201":
          description: routine creata con successo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutineOutput"
        "400":
          description: invalid input
        "500":
          description: errore interno del server

  /alerts:
    get:
      summary: recupera tutti gli Alert
      description: restituisce tutti gli Alert registrati
      operationId: getAllAlerts
      tags: [Alerts]
      responses:
        "200":
          description: lista di alert
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertOutput"
        "500":
          description: errore interno del server
    post:
      summary: crea un Alert
      description: crea un nuovo Alert contenente il body passato
      operationId: createAlert
      tags: [Alerts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
            example:
              oraInizio: "06:30"
              oraFine: "06:45"
              citta: "Silvi"
              tipoMessaggio: "Sveglia mattutina"
              codici:
                - 1001
      responses:
        "201":
          description: alert creato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertOutput"
        "400":
          description: invalid input
        "500":
          description: errore interno del server

  /alerts/{id}:
    get:
      summary: recupera un Alert
      description: restituisce un Alert univoco avente id passato
      operationId: getAlertById
      tags: [Alerts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID dell'alert
      responses:
        "200":
          description: alert trovato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertOutput"
        "404":
          description: alert not found
        "500":
          description: errore interno del server
    delete:
      summary: elimina un Alert
      description: elimina l'Alert avente l'id specificato
      operationId: deleteAlert
      tags: [Alerts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID dell'alert
      responses:
        "204":
          description: eliminazione avvenuta con successo
        "404":
          description: alert not found
        "500":
          description: errore interno del server
    patch:
      summary: patch Alert
      description: aggiorna parzialmente l'Alert avente id passato col body della request
      operationId: patchAlert
      tags: [Alerts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID dell'alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
            example:
              oraInizio: "06:35"
              oraFine: "06:50"
              citta: "Silvi"
              tipoMessaggio: "Sveglia aggiornata"
              codici:
                - 1003
      responses:
        "200":
          description: alert aggiornato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertOutput"
        "400":
          description: invalid input
        "404":
          description: alert not found
        "500":
          description: errore interno del server
    put:
      summary: replace Alert
      description: sostituisce completamente l'Alert avente id passato, se non esiste lo crea
      operationId: updateAlert
      tags: [Alerts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID dell'alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
            example:
              oraInizio: "06:30"
              oraFine: "06:45"
              citta: "Silvi"
              tipoMessaggio: "Sveglia mattutina"
              codici:
                - 1001
      responses:
        "200":
          description: alert sostituito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertOutput"
        "201":
          description: alert creato con successo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertOutput"
        "400":
          description: invalid input
        "500":
          description: errore interno del server

  /routines/{id}/alerts:
    get:
      summary: recupera gli Alert di una Routine
      description: recupera tutti gli Alert associati alla Routine avente id passato
      operationId: getAlertsForRoutine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID della routine
      tags: [associazione Alert Routine]
      responses:
        "200":
          description: lista di alert associati alla routine
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertOutput"
        "404":
          description: routine not found
        "500":
          description: errore interno del server
    post:
      summary: associa un Alert alla Routine
      description: Associa un alert esistente avente id passato nel body ad una routine specificata tramite il parametro id
      operationId: linkAlertToRoutine
      tags: [associazione Alert Routine]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID della routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alertId:
                  type: integer
                  format: int64
                  description: ID dell'alert
              required:
                - alertId
            example:
              alertId: 123
      responses:
        "201":
          description: alert associato con successo alla routine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertOutput"
        "400":
          description: invalid input
        "404":
          description: routine or alert not found
        "500":
          description: errore interno del server
    put:
      summary: crea un Alert e lo associa alla Routine
      description: crea una nuova risorsa Alert a partire dal body e la associa
        alla Routine identificata da id
      operationId: createAlertForRoutine
      tags: [associazione Alert Routine]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID della routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
            example:
              oraInizio: "06:30"
              oraFine: "06:45"
              citta: "Silvi"
              tipoMessaggio: "Sveglia mattutina"
              codici:
                - 1001
      responses:
        "201":
          description: alert creato e associato con successo alla routine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertOutput"
        "400":
          description: invalid input
        "404":
          description: routine not found
        "500":
          description: errore interno del server

  /routines/{id}/alerts/{alertId}:
    delete:
      summary: elimina un Alert da una Routine
      description: elimina l'associazione tra una Routine e un Alert specificati dai parametri id e alertId
      operationId: unlinkAlertFromRoutine
      tags: [associazione Alert Routine]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID della routine
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID dell'alert
      responses:
        "204":
          description: associazione rimossa
        "404":
          description: routine or alert not found
        "500":
          description: errore interno del server

components:
  schemas:
    RoutineInput:
      type: object
      properties:
        nomeRoutine:
          type: string
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/AlertInput"
      required:
        - nomeRoutine

    RoutineOutput:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nomeRoutine:
          type: string
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/AlertOutput"

    AlertInput:
      type: object
      properties:
        oraInizio:
          type: string
          example: "06:30"
        oraFine:
          type: string
          example: "06:45"
        citta:
          type: string
          example: "Silvi"
        tipoMessaggio:
          type: string
          example: "Sveglia mattutina"
        codici:
          type: array
          items:
            type: integer
            format: int64
      required:
        - oraInizio
        - oraFine
        - citta
        - tipoMessaggio
        - codici

    AlertOutput:
      type: object
      properties:
        id:
          type: integer
          format: int64
        oraInizio:
          type: string
          example: "06:30"
        oraFine:
          type: string
          example: "06:45"
        citta:
          type: string
          example: "Silvi"
        tipoMessaggio:
          type: string
          example: "Sveglia mattutina"
        codici:
          type: array
          items:
            type: integer
            format: int64